chore: examples optimization
