+++ examples/src/main/java/com/expediagroup/sdk/xap/examples/XapSdkDemoTestRun.java
@@ -17,9 +17,9 @@
 package com.expediagroup.sdk.xap.examples;
 
 import com.expediagroup.sdk.xap.examples.scenarios.lodging.AvailabilityCalendarsQuickStartScenario;
-import com.expediagroup.sdk.xap.examples.scenarios.lodging.ListingsHotelIdsSearchScenario;
+import com.expediagroup.sdk.xap.examples.scenarios.lodging.HotelIdsSearchEndToEndScenario;
 import com.expediagroup.sdk.xap.examples.scenarios.lodging.ListingsQuickStartScenario;
-import com.expediagroup.sdk.xap.examples.scenarios.lodging.QuotesQuickStartScenario;
+import com.expediagroup.sdk.xap.examples.scenarios.lodging.VrboPropertySearchEndToEndScenario;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -45,12 +45,13 @@ public class XapSdkDemoTestRun {
     ListingsQuickStartScenario listingsQuickStartScenario = new ListingsQuickStartScenario();
     listingsQuickStartScenario.run();
 
-    ListingsHotelIdsSearchScenario listingsHotelIdsSearchScenario =
-        new ListingsHotelIdsSearchScenario();
-    listingsHotelIdsSearchScenario.run();
+    HotelIdsSearchEndToEndScenario hotelIdsSearchEndToEndScenario =
+        new HotelIdsSearchEndToEndScenario();
+    hotelIdsSearchEndToEndScenario.run();
 
-    QuotesQuickStartScenario quotesQuickStartScenario = new QuotesQuickStartScenario();
-    quotesQuickStartScenario.run();
+    VrboPropertySearchEndToEndScenario vrboPropertySearchEndToEndScenario =
+        new VrboPropertySearchEndToEndScenario();
+    vrboPropertySearchEndToEndScenario.run();
 
     logger.info(
         "=============================== End of Lodging Scenarios ==============================");
