+++ examples/src/main/java/com/expediagroup/sdk/xap/examples/scenarios/lodging/HotelIdsSearchEndToEndScenario.java
@@ -1,3 +1,19 @@
+/*
+ * Copyright (C) 2024 Expedia, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 package com.expediagroup.sdk.xap.examples.scenarios.lodging;
 
 import com.expediagroup.sdk.core.model.Response;
@@ -20,30 +36,30 @@ import java.net.HttpURLConnection;
 import java.net.URL;
 import java.time.LocalDate;
 import java.util.ArrayList;
+import java.util.Collections;
 import java.util.HashMap;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
 import java.util.zip.ZipEntry;
 import java.util.zip.ZipInputStream;
+import org.apache.commons.lang3.StringUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 /**
- * This example demonstrates how to retrieve accessible property ids and location content from
- * SDP DownloadURL API and then get the prices of these properties using the Lodging Listings API.
+ * This example demonstrates how to retrieve accessible property ids from SDP DownloadURL API and
+ * then get the content and prices of these properties using the Lodging Listings API.
  *
- * <p>This is a common scenario for meta site partners. In practice, you can cache the property id
- * list along with content that does not change frequently (name, description, address, amenities,
- * etc.) to reduce heavy API calls and get only the prices of these properties in real-time from
- * the Lodging Listings API.
+ * <p>This is a common scenario for meta site partners. In practice, you can build a cache with the
+ * property id list, content and prices to improve respond time of your pages.
  */
-public class ListingsHotelIdsSearchScenario implements XapScenario {
+public class HotelIdsSearchEndToEndScenario implements XapScenario {
 
   private final XapClient client = createClient();
 
   private static final Logger LOGGER =
-      LoggerFactory.getLogger(ListingsHotelIdsSearchScenario.class);
+      LoggerFactory.getLogger(HotelIdsSearchEndToEndScenario.class);
 
   /**
    * This field limits the number of line to read from the SDP DownloadURL API Listings file to
@@ -54,29 +70,22 @@ public class ListingsHotelIdsSearchScenario implements XapScenario {
    */
   private static final int SAMPLE_ITEMS_RESTRICTION = 20;
 
-  /**
-   * A property id to location map. This mocks a cache in this example to store the static content
-   * of the properties.
-   */
-  private static final Map<String, String> PROPERTY_ID_AND_LOCATION_CACHE = new HashMap<>();
-
   public static void main(String[] args) {
-    new ListingsHotelIdsSearchScenario().run();
+    new HotelIdsSearchEndToEndScenario().run();
     System.exit(0);
   }
 
   @Override
   public void run() {
     LOGGER.info(
-        "======================== Running ListingsHotelIdsSearchScenario =======================");
+        "======================== Running HotelIdsSearchEndToEndScenario =======================");
 
     List<String> propertyIds = getPropertyIdsFromDownloadUrl();
-    cachePropertyLocationFromDownloadUrl(propertyIds);
-    HotelListingsResponse hotelListingsResponse = getPropertyPriceFromLodgingListings(propertyIds);
+    HotelListingsResponse hotelListingsResponse = getPropertiesFromLodgingListings(propertyIds);
     displayResult(hotelListingsResponse);
 
     LOGGER.info(
-        "========================== End ListingsHotelIdsSearchScenario =========================");
+        "========================== End HotelIdsSearchEndToEndScenario =========================");
   }
 
   /**
@@ -127,58 +136,23 @@ public class ListingsHotelIdsSearchScenario implements XapScenario {
   }
 
   /**
-   * Cache the location content from SDP DownloadURL API.
-   *
-   * @param propertyIds The property ids that need the location content.
-   */
-  private void cachePropertyLocationFromDownloadUrl(List<String> propertyIds) {
-    LOGGER.info(
-        "=============== Executing Step II: CachePropertyLocationFromDownloadUrl ================");
-    GetFeedDownloadUrlOperationParams getPropertyLocationParams =
-        GetFeedDownloadUrlOperationParams.builder()
-            // Use the type LOCATIONS to get the address of accessible properties.
-            .type(GetFeedDownloadUrlOperationParams.Type.LOCATIONS)
-            // Without any filters, this operation will return the information of all lodging
-            // properties in en_US by default.
-            .build();
-
-    Response<PresignedUrlResponse> downloadUrlLocationsResponse =
-        client.execute(new GetFeedDownloadUrlOperation(getPropertyLocationParams));
-
-    if (downloadUrlLocationsResponse.getData() == null
-        || downloadUrlLocationsResponse.getData().getBestMatchedFile() == null) {
-      throw new IllegalStateException("No location file found");
-    }
-
-    String locationsDownloadUrl = downloadUrlLocationsResponse.getData()
-        .getBestMatchedFile()
-        .getDownloadUrl();
-    LOGGER.info("Locations Download URL: {}", locationsDownloadUrl);
-
-    // Read and cache property locations from the file.
-    cachePropertyLocationFromLocationsFile(locationsDownloadUrl, propertyIds);
-
-    LOGGER.info(
-        "================ Step II: CachePropertyLocationFromDownloadUrl Executed ================");
-  }
-
-  /**
    * Get prices of the properties using the Lodging Listings API.
    *
    * @param propertyIds The property ids to get the prices.
    * @return The response of the Lodging Listings API.
    */
-  private HotelListingsResponse getPropertyPriceFromLodgingListings(List<String> propertyIds) {
+  private HotelListingsResponse getPropertiesFromLodgingListings(List<String> propertyIds) {
     LOGGER.info(
-        "================ Step III: Executing GetPropertyPriceFromLodgingListings ===============");
+        "================ Step II: Executing getPropertiesFromLodgingListings ===============");
 
     GetLodgingListingsOperationParams getLodgingListingsOperationParams =
         GetLodgingListingsOperationParams.builder()
             .partnerTransactionId(PARTNER_TRANSACTION_ID)
             // Use the property ids read from the file
             .ecomHotelIds(new HashSet<>(propertyIds))
-            // Use LOWEST to get minimal content details to reduce the response size
-            .contentDetails(GetLodgingListingsOperationParams.ContentDetails.LOWEST)
+            // The links to return, WEB includes WS (Web Search Result Page)
+            // and WD (Web Details Page)
+            .links(Collections.singletonList(GetLodgingListingsOperationParams.Links.WEB))
             // Check-in 5 days from now
             .checkIn(LocalDate.now().plusDays(5))
             // Check-out 10 days from now
@@ -193,7 +167,7 @@ public class ListingsHotelIdsSearchScenario implements XapScenario {
             .getData();
 
     LOGGER.info(
-        "================ Step III: GetPropertyPriceFromLodgingListings Executed ================");
+        "================ Step II: getPropertiesFromLodgingListings Executed ================");
     return hotelListingsResponse;
   }
 
@@ -283,109 +257,12 @@ public class ListingsHotelIdsSearchScenario implements XapScenario {
   }
 
   /**
-   * Caches the location content of the properties from the file pointed by the download URL.
-   *
-   * @param locationsDownloadUrl The download URL of the zip file containing the property locations.
-   * @param propertyIds          The property ids to get the location content.
-   */
-  private void cachePropertyLocationFromLocationsFile(String locationsDownloadUrl,
-                                                      List<String> propertyIds) {
-    HttpURLConnection connection = null;
-    try {
-      // Open a connection to the URL
-      URL url = new URL(locationsDownloadUrl);
-      connection = (HttpURLConnection) url.openConnection();
-      connection.setRequestMethod("GET");
-      connection.setDoInput(true);
-
-      try (ZipInputStream zipStream = new ZipInputStream(connection.getInputStream())) {
-        ZipEntry entry;
-        while ((entry = zipStream.getNextEntry()) != null) {
-          if (entry.getName().endsWith(".jsonl")) {
-            LOGGER.info("Reading property locations from file: {}", entry.getName());
-            try (BufferedReader reader = new BufferedReader(new InputStreamReader(zipStream))) {
-              String line;
-              ObjectMapper objectMapper = new ObjectMapper();
-              while ((line = reader.readLine()) != null
-                  && PROPERTY_ID_AND_LOCATION_CACHE.size() < propertyIds.size()) {
-                // Parse the property location from the json object
-                // An example json line from the jsonl file:
-                /*
-                {
-                  "propertyId": {
-                    "expedia": "1234567",
-                    "hcom": "123456789",
-                    "vrbo": "123.1234567.7654321"
-                  },
-                  "propertyType": {
-                    "id": 16,
-                    "name": "Apartment"
-                  },
-                  "propertyName": "Expedia Property Name",
-                  "address1": "",
-                  "address2": "",
-                  "city": "Bains-les-Bains",
-                  "province": "Grand Est",
-                  "country": "France",
-                  "postalCode": "88240",
-                  "geoLocation": {
-                    "latitude": "10.999999",
-                    "longitude": "-10.999999",
-                    "obfuscated": false
-                  },
-                  "locationAttribute": {
-                    "city": {
-                      "id": "553248635975861971",
-                      "name": "Forges-les-Bains"
-                    },
-                    "region": {
-                      "id": "553248634539665703",
-                      "name": "Paroisse Saint-Colomban-en-Vôge"
-                    },
-                    "distanceFromCityCenter": {
-                      "distance": "0.21",
-                      "unit": "km"
-                    }
-                  }
-                }
-                */
-                JsonNode jsonNode = objectMapper.readTree(line);
-                // Check if the property id is in the list
-                if (propertyIds.contains(jsonNode.get("propertyId").get("expedia").asText())) {
-                  // Get the location content of the property
-                  String location = jsonNode.get("propertyName").asText() + ", "
-                      + jsonNode.get("city").asText() + ", "
-                      + jsonNode.get("province").asText() + ", "
-                      + jsonNode.get("country").asText();
-                  // Store the location content in the cache
-                  PROPERTY_ID_AND_LOCATION_CACHE.put(
-                      jsonNode.get("propertyId")
-                          .get("expedia")
-                          .asText(),
-                      location);
-                }
-              }
-            }
-          }
-        }
-      }
-
-    } catch (IOException e) {
-      LOGGER.error("Error reading property locations from download URL", e);
-    } finally {
-      if (connection != null) {
-        connection.disconnect();
-      }
-    }
-  }
-
-  /**
    * Display the result of the operations.
    *
    * @param hotelListingsResponse The response of the Lodging Listings API.
    */
   private static void displayResult(HotelListingsResponse hotelListingsResponse) {
-    LOGGER.info("======================= Executing Step IV: DisplayResult =======================");
+    LOGGER.info("====================== Executing Step III: DisplayResult =======================");
     if (hotelListingsResponse == null || hotelListingsResponse.getHotels() == null
         || hotelListingsResponse.getHotels().isEmpty()) {
       throw new IllegalStateException("No properties found.");
@@ -403,15 +280,38 @@ public class ListingsHotelIdsSearchScenario implements XapScenario {
       }
       LOGGER.info(
           "=================================== Property Start ===================================");
-      String hotelId = hotel.getId();
-
-      // Get the location content of the property from the cache
-      LOGGER.info("Property Id: {}", hotelId);
-      LOGGER.info("Cached Property Location: {}", PROPERTY_ID_AND_LOCATION_CACHE.get(hotelId));
-
-      // Get the price of the property from the room type
+      // To get the property name
+      if (StringUtils.isNotEmpty(hotel.getName())) {
+        LOGGER.info("Property Name: {}", hotel.getName());
+      }
+      // To get the property address
+      if (hotel.getLocation() != null) {
+        LOGGER.info("Property Address: {}", hotel.getLocation().getAddress());
+      }
+      // To get the property thumbnail URL
+      if (StringUtils.isNotEmpty(hotel.getThumbnailUrl())) {
+        LOGGER.info("Thumbnail URL: {}", hotel.getThumbnailUrl());
+      }
+      // To get the star rating of the property. The value is between 1.0 and 5.0
+      // in a 0.5 increment.
+      if (hotel.getStarRating() != null) {
+        LOGGER.info("Star Rating: {}", hotel.getStarRating().getValue());
+      }
+      // To get the guest rating of the property. The value is between 1.0 and 5.0
+      // in a 0.1 increment.
+      if (StringUtils.isNotEmpty(hotel.getGuestRating())) {
+        LOGGER.info("Guest Rating: {}", hotel.getGuestRating());
+      }
+      // To get the total number of reviews for the property
+      if (hotel.getGuestReviewCount() != null) {
+        LOGGER.info("Review Count: {}", hotel.getGuestReviewCount());
+      }
       if (hotel.getRoomTypes() != null && !hotel.getRoomTypes().isEmpty()) {
+        // To get the first room type information
         RoomType roomType = hotel.getRoomTypes().get(0);
+        if (StringUtils.isNotEmpty(roomType.getDescription())) {
+          LOGGER.info("Room Type: {}", roomType.getDescription());
+        }
         if (roomType.getPrice() != null) {
           // To get the total price of the room type
           if (roomType.getPrice().getTotalPrice() != null) {
@@ -426,10 +326,27 @@ public class ListingsHotelIdsSearchScenario implements XapScenario {
                 roomType.getPrice().getAvgNightlyRate().getCurrency());
           }
         }
+        // To get the free cancellation flag of the selected room
+        if (roomType.getRatePlans() != null && !roomType.getRatePlans().isEmpty()
+            && roomType.getRatePlans().get(0).getCancellationPolicy() != null) {
+          LOGGER.info("Free Cancellation: {}",
+              roomType.getRatePlans().get(0).getCancellationPolicy().getFreeCancellation());
+        }
+        if (roomType.getLinks() != null) {
+          // To get the deeplink to the Expedia Web Search Result Page
+          if (roomType.getLinks().getWebSearchResult() != null) {
+            LOGGER.info("WebSearchResult Link: {}",
+                roomType.getLinks().getWebSearchResult().getHref());
+          }
+          // To get the deeplink to the Expedia Web Details Page
+          if (roomType.getLinks().getWebDetails() != null) {
+            LOGGER.info("WebDetails Link: {}", roomType.getLinks().getWebDetails().getHref());
+          }
+        }
       }
       LOGGER.info(
           "==================================== Property End ====================================");
     });
-    LOGGER.info("======================= Step IV: DisplayResult Executed ========================");
+    LOGGER.info("====================== Step III: DisplayResult Executed ========================");
   }
 }
